requirements.txt
flask==3.1.0
langchain==0.3.23
openai==1.75.0
python-dotenv==1.1.0
